Forget the whole optional logger thing for parsers, for now.
If this becomes actually-useful at some point, we'll bring it back.
But really, we should be intelligently setting up Logger as the common
point of reference, not optionally shoving things in for tools that
didn't work with the default output settings for Logger.
We should make it easier to stick application loggers into the Logger
class for use by lower-level stuff.