Rethink the type-by-type parsing in conversions class.
Instead of a bunch of private parse functions that have debug printouts
if they didn't recognize the type of the value arg,
we can have those private ones just attempt to parse from types they
recognize, and return None if it's unrecognized.
Then, we have public-facing "To<Type>" functions, which will always
return the designated type, and do the printing of a warning message
before trying to coerce an unrecognized type,
e.g. by naively calling type constructor on the value.

This will enable the ConvertToType class to better handle the case where
a list of candidate types is given.