Add a 'soft' parsing function that returns None if the type of value wasn't directly handled for given to_type.
This is then used in ConvertToType for the case where to_type is a list.
First, it looks for a direct match between type of value and a type in
the list.
If none is found, it looks for a type in the list that can handle the
type of value.
If none is found, it falls back on 'hard' parsing with first type in the
list, which will force the value into a known type, like str, and parse
from that type.