Fix params given for to_type in testing calls, and add a few more test cases.
Truthfully, most of these are truly tests of the conversion function,
under the hood.
Like, they're not directly testing behvior of the ParseElement function.
Still, they're testing different use cases of the function itself,
since it's mostly a convenient wrapper around the underlying conversion
function, so... guess we'll keep them for now.
Not really sure of the underlying best practice for testing of wrapper
functions, in any case we need these to exist somewhere.