Update TableSchema functioning after new work on location schemas.
In particular, make it look for a specifier of the type of location in
whatever config is being read, and load the specific type of location
schema subclass.
In theory, could use some sort of factory approach for this in the
future.

Also, update the _parseStructure to allow config to specify a file with
the actual structure schema config, rather than requiring the dict we
set up from to contain the whole spec directly.