Have ParseElement track original keys of given dict.
We use a local dict that uses upper-cased keys, which map to a tuple of
the original value and the original key.
Then when we find a match, we can still easily retrieve the original
value, but also easily get the original key, and use it for the del
issue. This addresses the test breakages that lead to filing issue #127.