Officially retire the general-purpose 'SELECT' assembler function.
It was pretty cool how 5 or 6 years ago, we had a pretty-comprehensive
function for creating SELECT queries from a bunch of parameters.
However, in practice we tended to constrain our queries to just a couple
specific tasks, so we never really needed its general-purpose powers.

Maybe deserves an entry in the 'historical bits' part of the
documentation project.