Don't include http user agent in built-ins.
We haven't ever used it, and it technically can be used for
fingerprinting.
I think at one point, idea was to use it to better understand how people
were accessing our games, but in practice we haven't been using it, and
better to not include it for now, in that case.