Merge _name params into related param.
We accept a string or appropriate Config/Schema as inputs to
constructor, and figure out what to do based on what was given.
Simplifies the interface, and ensures that *if* someone passes in the
objects wholly-formed, they're not in a weird spot of choosing what name
to pass with it, or wondering whether a name is necessary or not.

Also, added a data_stores param, totally optional dict for looking up a
DataStoreConfig by name, in case we ever find a way to shift that around
for use with initialization of stuff in core. Theoretically, we could
have the FromDict use the constructor directly, passing None for
everything that needs to be parsed, but passing in the pre-parsed
data_stores dict. That actually should work. Will plan on it.