Rename REMOTE_URL -> PUBLIC_URL.
Generally, this is going to represent where public can access a dataset
repo. So call it public, rather than remote, which implies we're
configuring some thing that is remote and local all at once.
Then interfaces can look for a public URL in config to access data from,
if the repository doesn't have a local path.