Have FileStoreConfig manage extension and checking for zipped file.
Useful for any connectors that are going to connect to a single file,
such as CSVConnector, so they don't need extra params.
They can just check if the given config points at a CSV/TSV, and/or points
at a zip.
Then config is handled by config, and they just behave according to
config instead of checking for consistency between config and some
additional params.